fix: msising schema extension Zeiteinheit PERCENTAGE
It is not enough to only adjust Zeiteinheit type! Each Avro DTO contains
the entire schema as JSON. Even when the Java classes are extended,
the deserialization fails when an unknown value in the JSON schema is
encountered.

We've had the issue that a BetreuungAnfrageDTO was stored in Kafka with
PERCENTAGE, but the kiBon consumer (BetreuungAnfrageDTO) did not contain
the PERCENTAGE in it's schema. Thus, the deserializer returned 2 as
enum value and crashed with an IndexOutOfBound exception.